Extract VariableReferenceCounter and StackTooDeep error from EVMCodeTransform.
