+++ libyul/Scope.cpp
@@ -21,6 +21,8 @@
 
 #include <libyul/Scope.h>
 
+#include <libsolutil/Visitor.h>
+
 using namespace std;
 using namespace solidity;
 using namespace solidity::yul;
